+++ test/unit/specs/create-matcher.spec.js
@@ -22,6 +22,57 @@ describe('Creating Matcher', function () {
     process.env.NODE_ENV = 'production'
   })
 
+  it('can add nested routes', function () {
+    const component = { name: 'fake' }
+    const matcher = createMatcher([
+      {
+        path: '/p',
+        name: 'parent',
+        children: [
+          { path: 'a', name: 'a', component },
+          {
+            path: 'c',
+            name: 'child',
+            component,
+            children: [{ path: 'n', name: 'nested', component }]
+          }
+        ]
+      },
+      {
+        // easier to debug tests
+        path: '*', name: 'not-found', component
+      }
+    ])
+
+    matcher.addRoute({ path: '/b', name: 'b', component })
+    matcher.addRoute('parent', { path: 'b', name: 'p-b', component })
+    matcher.addRoute('child', { path: 'b', name: 'p-c-b', component })
+
+    expect(matcher.match('/b').name).toBe('b')
+    expect(matcher.match('/p/b').name).toBe('p-b')
+    expect(matcher.match('/p/c/b').name).toBe('p-c-b')
+    expect(matcher.match('/p/c/n').name).toBe('nested')
+  })
+
+  it('can get all routes', function () {
+    const component = { name: 'fake' }
+    const matcher = createMatcher([])
+
+    expect(matcher.getRoutes()).toEqual([])
+
+    matcher.addRoute({ path: '/b', name: 'b', component })
+    expect(matcher.getRoutes().length).toBe(1)
+
+    matcher.addRoute({ path: '/c', name: 'c', alias: ['/a', '/d'], component })
+    expect(matcher.getRoutes().length).toBe(4)
+
+    matcher.addRoute('b', { path: 'd', component })
+    expect(matcher.getRoutes().length).toBe(5)
+
+    matcher.addRoute('c', { path: 'd', component })
+    expect(matcher.getRoutes().length).toBe(8)
+  })
+
   it('in development, has logged a warning if a named route does not exist', function () {
     process.env.NODE_ENV = 'development'
     const { name, matched } = match({ name: 'bar' }, routes[0])