+++ types/router.d.ts
@@ -47,6 +47,11 @@ export declare class VueRouter {
   onReady(cb: Function, errorCb?: ErrorHandler): void
   onError(cb: ErrorHandler): void
   addRoutes(routes: RouteConfig[]): void
+
+  addRoute(parent: string, route: RouteConfig): void
+  addRoute(route: RouteConfig): void
+  getRoutes(): RouteRecordPublic[]
+
   resolve(
     to: RawLocation,
     current?: Route,
@@ -70,6 +75,8 @@ export declare class VueRouter {
   static NavigationFailureType: {
     [k in keyof typeof NavigationFailureType]: NavigationFailureType
   }
+
+  static START_LOCATION: Route
 }
 
 export enum NavigationFailureType {
@@ -157,6 +164,26 @@ export interface RouteRecord {
     | RoutePropsFunction
     | Dictionary<boolean | Object | RoutePropsFunction>
 }
+
+export interface RouteRecordPublic {
+  path: string
+  components: Dictionary<Component>
+  instances: Dictionary<Vue>
+  name?: string
+  redirect?: RedirectOption
+  meta: any
+  beforeEnter?: (
+    route: Route,
+    redirect: (location: RawLocation) => void,
+    next: () => void
+  ) => any
+  props:
+    | boolean
+    | Object
+    | RoutePropsFunction
+    | Dictionary<boolean | Object | RoutePropsFunction>
+}
+
 
 export interface Location {
   name?: string